clean-up of imports, merging two seaicedeformation files
