+++ src/SeaIceDeformation/utils_compute_deformations.py
@@ -54,18 +54,18 @@ single output netcdf4 file that combines all processed datasets.
 
 '''
 
+# Loading from default packages
 import csv
 import datetime
 import os
 from math import sqrt
-
 import numpy as np
 from netCDF4 import Dataset
-
-import utils_datetime
-import utils_deformation_computations as deformation_comp
-import utils_load_data as load_data
 from tqdm import tqdm
+
+# Loading from other files
+import SeaIceDeformation.utils_datetime as utils_datetime
+import SeaIceDeformation.utils_load_data as load_data
 
 def compute_deformations(config=None):
     '''
@@ -189,10 +189,10 @@ def compute_deformations(config=None):
             no.append(file_num)
 
             # Create a list of velocity components for each triangle vertex
-            u_list, v_list = deformation_comp.calculate_uv_lists( sx_list, ex_list, sy_list, ey_list, dt)
+            u_list, v_list = calculate_uv_lists( sx_list, ex_list, sy_list, ey_list, dt)
 
             # Compute the strain rates
-            dudx, dudy, dvdx, dvdy = deformation_comp.calculate_strainRates( u_list, v_list, sx_list, sy_list )
+            dudx, dudy, dvdx, dvdy = calculate_strainRates( u_list, v_list, sx_list, sy_list )
 
             # Compute the divergence rate
             eps_I = dudx + dvdy
@@ -402,6 +402,77 @@ def compute_deformations(config=None):
 
     # # Close dataset
     # output_ds.close()
+
+
+
+def calculate_uv_lists( sx_list, ex_list, sy_list, ey_list, dT):
+    ''' (list, list, float) -> float(list, list)
+
+    Function that calculates u and v velocity components of
+    each cell vertices and returns them as a list.
+
+    Returns a tuple (u_list, v_list) of the lists of velocity components.
+
+    Keyword arguments: \\
+    sx_list -- list of starting x positions of each cell vertices \\
+    ex_list -- list of ending x positions of each cell vertices \\
+    sy_list -- list of starting x positions of each cell vertices \\
+    ey_list -- list of ending x positions of each cell vertices \\
+    dT      -- time interval
+    '''
+
+    # Compute the u and v velocity components at the current vertex
+    u_list = (ex_list - sx_list) / dT
+    v_list = (ey_list - sy_list) / dT
+
+    return u_list, v_list
+
+
+def calculate_strainRates( u_list, v_list, sx_list, sy_list ):
+    ''' (list, list, list, list) -> tuple(float, float, float, float)
+
+    Computes the strain rates (or velocity derivatives).
+
+    Keyword arguments: \\
+    u_list -- list of u component velocities for each cell vertices \\
+    v_list -- list of v component velocities for each cell vertices \\
+    sx_list -- list of starting x positions for each cell vertices \\
+    sy_list -- list of starting y positions for each cell vertices \\
+    '''
+
+    # Find the number of cell vertices
+    n = len(sx_list)
+
+    #----- Compute the Lagrangian cell area A -----
+
+    # Initialize the cell area to 0
+    A = 0.0
+
+    # Perform a summation to compute A (see Bouchat et al. (2020) eqn. 6)
+    for i in range( n ):
+        A += (1.0/2.0) * ( sx_list[i] * sy_list[((i+1) % n)] - sx_list[((i+1) % n)] *  sy_list[i])
+
+    #----- Compute the strain rates ---------------
+
+    # Initialize the strain rates to 0
+    dudx = 0.0
+    dudy = 0.0
+    dvdx = 0.0
+    dvdy = 0.0
+
+    # Perform a summation to compute strain rates (see Bouchat et al. (2020) eqn. 5)
+    for i in range( n ):
+
+        dudx += 1.0/(2.0*A)  * ( u_list[((i+1) % n)] + u_list[i] ) * ( sy_list[((i+1) % n)] - sy_list[i] )
+
+        dudy += -1.0/(2.0*A) * ( u_list[((i+1) % n)] + u_list[i] ) * ( sx_list[((i+1) % n)] - sx_list[i] )
+
+        dvdx += 1.0/(2.0*A)  * ( v_list[((i+1) % n)] + v_list[i] ) * ( sy_list[((i+1) % n)] - sy_list[i] )
+
+        dvdy += -1.0/(2.0*A) * ( v_list[((i+1) % n)] + v_list[i] ) * ( sx_list[((i+1) % n)] - sx_list[i] )
+
+    return dudx, dudy, dvdx, dvdy
+
 
 if __name__ == '__main__':
     compute_deformations()
