+++ src/SeaIceDeformation/utils_grid_coord_system.py
@@ -9,6 +9,7 @@ Helper code for the conversion of a triangle's lat,lon coordinates to
 x,y coordinates in a local cartesian grid coordinate system.
 '''
 
+# Loading from default packages
 import numpy as np
 from haversine import haversine
 from math import sqrt, acos
