+++ src/SeaIceDeformation/utils_load_data.py
@@ -8,8 +8,8 @@ Utils - Load data
 Helper code for loading files from all stages of data processing.
 '''
 
+# Loading from default packages
 import os
-
 import pandas as pd
 
 # Create a class of errors for csv loading and two subclasses
